move validation to src from model
