+++ wasm-rpc-stubgen/src/model/mod.rs
@@ -1,4 +1,3 @@
 pub mod oam;
 pub mod unknown_properties;
-pub mod validation;
 pub mod wasm_rpc;
