+++ wasm-rpc-stubgen/src/validation.rs
@@ -1,7 +1,7 @@
 use itertools::Itertools;
 use std::fmt::Display;
 
-// TODO: move to common
+
 
 pub struct ValidationContext {
     pub name: &'static str,
